// rce/src/stub/logger_pb.ts
export enum Level {
    /**
     * @generated from protobuf enum value: DEBUG = 0;
     */
    DEBUG = 0,
    /**
     * @generated from protobuf enum value: INFO = 1;
     */
    INFO = 1,
    /**
     * @generated from protobuf enum value: WARNING = 2;
     */
    WARNING = 2,
    /**
     * @generated from protobuf enum value: ERROR = 3;
     */
    ERROR = 3,
    /**
     * @generated from protobuf enum value: CRITICAL = 4;
     */
    CRITICAL = 4
}